Refactor policy.ts to reflect the workflow interface.
