+++ cli/src/lib/policy/context.ts
@@ -1,10 +1,9 @@
-export interface Result {
-  resource: string;
-  message: string;
-}
+import {Resource, Result} from './policy'
 
 export class Context {
   readonly inputs: object
+
+  private resources: Map<string, Resource[]> = new Map<string, Resource[]>()
 
   private results: Map<string, Result[]> = new Map<string, Result[]>()
 
@@ -25,5 +24,16 @@ export class Context {
       this.results.set(ruleKey, results)
     }
     this.results.get(ruleKey)?.push(...results)
+  }
+
+  reportResource(resources: Resource[], providerKey: string): void {
+    if (!this.resources.has(providerKey)) {
+      this.resources.set(providerKey, resources)
+    }
+    this.resources.get(providerKey)?.push(...resources)
+  }
+
+  getProviderResources(providerKey: string): Array<Resource> | undefined {
+    return this.resources.get(providerKey)
   }
 }
