+++ cli/src/lib/policy/policy.ts
@@ -1,18 +1,44 @@
-import {Context, Result} from './context'
+import {Context} from './context'
 import * as path from 'path'
 import * as fs from 'fs'
 import * as _ from 'lodash'
 import YamlReader from '../shared/YamlReader'
-import {PolicySpec} from './spec'
+import {PolicySpec, RuleSpec} from './spec'
 import {PolicyConfig} from '../config'
 
-const DEFAULT_POLICY_SPEC_FILE_NAME = 'stacklint-workflow.yaml'
+const DEFAULT_POLICY_SPEC_FILE_NAME = 'stacklint-policy.yaml'
 
 export interface Resource {
   [key: string]:
     {
       type: string;
       properties: { [key: string]: any };
+    };
+}
+
+export interface Result {
+  resource: string;
+  message: string;
+}
+
+export interface PolicyInfo {
+  PolicyConfig : PolicyConfig;
+  PolicySpec: PolicySpec;
+}
+
+export interface CheckingPlan {
+  resources: {
+    [key: string]: Resource;
+  };
+  rules: {
+    [key: string]: RuleSpec;
+  };
+}
+
+export interface FixingPlan {
+  [key: string]:
+    {
+      type: string;
     };
 }
 
@@ -27,8 +53,6 @@ export class Policy {
 
   private readonly context: Context
 
-  private resources: Map<string, Resource[]> = new Map<string, Resource[]>()
-
   private constructor(policyConfig: PolicyConfig, policyName: string, policyPath: string, policySpec: PolicySpec) {
     this.policyName = policyName
     this.policyConfig = policyConfig
@@ -37,7 +61,7 @@ export class Policy {
     this.context = new Context(policyConfig)
   }
 
-  static getPolicySpec(policyPath: string): PolicySpec {
+  static loadPolicySpec(policyPath: string): PolicySpec {
     return YamlReader.load(policyPath + path.sep + DEFAULT_POLICY_SPEC_FILE_NAME)
   }
 
@@ -53,10 +77,61 @@ export class Policy {
     } else {
       throw new Error('Node module based policy sharing is not supported yet')
     }
-    return new Policy(policyConfig, name, policyPath, this.getPolicySpec(policyPath))
+    return new Policy(policyConfig, name, policyPath, this.loadPolicySpec(policyPath))
+  }
+
+  async show(): Promise<PolicyInfo> {
+
+    return null
+  }
+
+  async plan(): Promise<CheckingPlan> {
+    const results: Resource[] = new Array<Resource>()
+    await Promise.all(Object.keys(this.policySpec.providers).map(async providerKey => {
+      const res = await this.getProviderResources(providerKey)
+      if (res !== undefined) {
+        results.push(...res)
+      }
+    }))
+    return results
+  }
+
+  async check(checkingPlan?: CheckingPlan): Promise<FixingPlan | null> {
+    /**
+     await Promise.all(Object.keys(this.stackLintConfig.plugins).map(async key => {
+      const plugin = await Policy.build(this.stackLintConfig.plugins[key])
+      const res = await plugin.getAllRuleResults()
+      results.push(...res)
+    }))
+     */
+    return null
+  }
+
+  async fix(fixingPlan?: FixingPlan): Promise<boolean> {
+    /**
+     await Promise.all(Object.keys(this.stackLintConfig.plugins).map(async key => {
+      const plugin = await Policy.build(this.stackLintConfig.plugins[key])
+      const res = await plugin.getAllRuleResults()
+      results.push(...res)
+    }))
+     */
+    return true
   }
 
   /**
+   private async getResourcesForProvider(providerKey: string): Promise<Resource[]> {
+    if (_.has(this.policySpec.providers, providerKey)) {
+      if (this.context.getProviderResources(providerKey) === undefined) {
+        const provider = this.policySpec.providers[providerKey]
+        const providerPath = path.resolve(this.policyPath, provider.main)
+        const providerMain = require(providerPath)
+        providerMain.handler(this.context)
+      }
+      return this.context.getProviderResources(providerKey)
+    }
+    throw new Error('Can not find the provider')
+  }
+
    async getAllRuleResults(): Promise<Result[]> {
     const results: Result[] = new Array<Result>()
     const resources = await this.getAllResources()
@@ -79,41 +154,5 @@ export class Policy {
     }
     throw new Error('can not find the provider')
   }
-
    */
-
-  async getAllResources(): Promise<Resource[]> {
-    const results: Resource[] = new Array<Resource>()
-    await Promise.all(Object.keys(this.policySpec.providers).map(async providerKey => {
-      const res = await this.getProviderResources(providerKey)
-      if (res !== undefined) {
-        results.push(...res)
-      }
-    }))
-    return results
-  }
-
-  private async getProviderResources(providerKey: string): Promise<Resource[]> {
-    if (_.has(this.policySpec.providers, providerKey)) {
-      if (this.context.getProviderResources(providerKey) === undefined) {
-        const provider = this.policySpec.providers[providerKey]
-        const providerPath = path.resolve(this.policyPath, provider.main)
-        const providerMain = require(providerPath)
-        providerMain.handler(this.context)
-      }
-      return this.context.getProviderResources(providerKey)
-    }
-    throw new Error('Can not find the provider')
-  }
-
-  reportResource(resources: Resource[], providerKey: string): void {
-    if (!this.resources.has(providerKey)) {
-      this.resources.set(providerKey, resources)
-    }
-    this.resources.get(providerKey)?.push(...resources)
-  }
-
-  getProviderResources(providerKey: string): Array<Resource> | undefined {
-    return this.resources.get(providerKey)
-  }
 }
